minerador/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
schedule==1.2.0
python-dotenv==1.0.0
lxml==4.9.3
urllib3==2.1.0
fake-useragent==1.4.0
zipfile36==0.1.3
xlrd==2.0.1
psutil==5.9.6
